+++ build-tools/packages/build-tools/src/common/biomeConfig.ts
@@ -12,10 +12,8 @@ import { merge } from "ts-deepmerge";
 // Opaque is deprecated in newer type-fest versions (replaced by Tagged); we pin type-fest 2.x for current CommonJS
 import type { Opaque } from "type-fest";
 
-import { Biome2ConfigReader } from "./biome2Config";
 import type { Configuration as BiomeConfigRaw } from "./biomeConfigTypes";
 import { getClosestBiomeConfigPath, loadRawBiomeConfigFile } from "./biomeConfigUtils";
-import { type BiomeMajorVersion, detectBiomeVersion } from "./biomeVersion";
 import type { GitRepo } from "./gitRepo";
 
 /**
@@ -245,62 +243,4 @@ export class BiomeConfigReaderV1 {
 		const files = await getBiomeFormattedFiles(mergedConfig, directory, gitRepo);
 		return new BiomeConfigReaderV1(configFile, allConfigs, mergedConfig, files);
 	}
-}
-
-/**
- * A common interface for both Biome 1.x and 2.x config readers.
- * This interface defines the properties that are available on both readers.
- */
-export interface BiomeConfigReader {
-	/**
-	 * The absolute path to the closest (most specific) config file.
-	 */
-	readonly closestConfig: string;
-	/**
-	 * The directory containing the config file.
-	 */
-	readonly directory: string;
-	/**
-	 * All config file paths, in order of application (base configs first).
-	 */
-	readonly allConfigs: string[];
-	/**
-	 * Absolute paths to files that would be formatted by Biome.
-	 */
-	readonly formattedFiles: string[];
-}
-
-/**
- * Creates the appropriate BiomeConfigReader based on the detected Biome version.
- *
- * This factory function auto-detects whether Biome 1.x or 2.x is installed and returns
- * the appropriate config reader. Use this function when you want automatic version detection.
- *
- * @param directoryOrConfigFile - A path to a directory or a Biome config file.
- * @param gitRepo - A GitRepo instance that is used to enumerate files.
- * @param forceVersion - If provided, forces the use of a specific Biome version reader
- *                       instead of auto-detecting.
- * @returns A BiomeConfigReader (for 1.x) or Biome2ConfigReader (for 2.x) based on the detected version.
- */
-export async function createBiomeConfigReader(
-	directoryOrConfigFile: string,
-	gitRepo: GitRepo,
-	forceVersion?: BiomeMajorVersion,
-): Promise<BiomeConfigReader> {
-	let majorVersion: BiomeMajorVersion;
-
-	if (forceVersion !== undefined) {
-		majorVersion = forceVersion;
-	} else {
-		// Auto-detect the Biome version
-		const versionInfo = await detectBiomeVersion(directoryOrConfigFile);
-		majorVersion = versionInfo?.majorVersion ?? 1; // Default to 1.x if detection fails
-	}
-
-	if (majorVersion === 2) {
-		return Biome2ConfigReader.create(directoryOrConfigFile, gitRepo);
-	}
-
-	// Default to Biome 1.x reader
-	return BiomeConfigReaderV1.create(directoryOrConfigFile, gitRepo);
 }
