+++ build-tools/packages/build-tools/src/common/biomeConfigUtils.ts
@@ -10,6 +10,8 @@
 import { readFile } from "node:fs/promises";
 import * as JSON5 from "json5";
 
+import type { GitRepo } from "./gitRepo";
+
 // switch to regular import once building ESM
 const findUp = import("find-up");
 
@@ -57,4 +59,67 @@ export async function getClosestBiomeConfigPath(
 			}
 			return config;
 		});
+}
+
+/**
+ * A common interface for both Biome 1.x and 2.x config readers.
+ * This interface defines the properties that are available on both readers.
+ */
+export interface BiomeConfigReader {
+	/**
+	 * The absolute path to the closest (most specific) config file.
+	 */
+	readonly closestConfig: string;
+	/**
+	 * The directory containing the config file.
+	 */
+	readonly directory: string;
+	/**
+	 * All config file paths, in order of application (base configs first).
+	 */
+	readonly allConfigs: string[];
+	/**
+	 * Absolute paths to files that would be formatted by Biome.
+	 */
+	readonly formattedFiles: string[];
+}
+
+/**
+ * Creates the appropriate BiomeConfigReader based on the detected Biome version.
+ *
+ * This factory function auto-detects whether Biome 1.x or 2.x is installed and returns
+ * the appropriate config reader. Use this function when you want automatic version detection.
+ *
+ * @param directoryOrConfigFile - A path to a directory or a Biome config file.
+ * @param gitRepo - A GitRepo instance that is used to enumerate files.
+ * @param forceVersion - If provided, forces the use of a specific Biome version reader
+ *                       instead of auto-detecting.
+ * @returns A BiomeConfigReader (for 1.x) or Biome2ConfigReader (for 2.x) based on the detected version.
+ */
+export async function createBiomeConfigReader(
+	directoryOrConfigFile: string,
+	gitRepo: GitRepo,
+	forceVersion?: 1 | 2,
+): Promise<BiomeConfigReader> {
+	// Import lazily to avoid circular dependencies
+	const { BiomeConfigReaderV1 } = await import("./biomeConfig.js");
+	const { Biome2ConfigReader } = await import("./biome2Config.js");
+	const { detectBiomeVersion } = await import("./biomeVersion.js");
+
+	let majorVersion: 1 | 2;
+
+	if (forceVersion !== undefined) {
+		majorVersion = forceVersion;
+	} else {
+		// Auto-detect the Biome version
+		const versionInfo = await detectBiomeVersion(directoryOrConfigFile);
+		majorVersion = versionInfo?.majorVersion ?? 1; // Default to 1.x if detection fails
+	}
+
+	if (majorVersion === 2) {
+		return Biome2ConfigReader.create(directoryOrConfigFile, gitRepo);
+	}
+
+	// Default to Biome 1.x reader
+	return BiomeConfigReaderV1.create(directoryOrConfigFile, gitRepo);
 }
